+++ modules/vectorStyle/pointStyle.js
@@ -1,5 +1,5 @@
 import StyleModel from "./style.js";
-import {Circle as CircleStyle, Fill, Stroke, Style, Icon} from "ol/style.js";
+import {Circle as CircleStyle, Fill, Stroke, Style, Icon, RegularShape} from "ol/style.js";
 import {prepareValue, isObjectPath} from "../../src/utils/attributeMapper.js";
 
 const PointStyleModel = StyleModel.extend(/** @lends PointStyleModel.prototype */{
@@ -69,7 +69,16 @@ const PointStyleModel = StyleModel.extend(/** @lends PointStyleModel.prototype *
         "circleBarCircleStrokeWidth": 1,
         "circleBarLineStrokeColor": [0, 0, 0, 1],
         "scalingAttribute": "",
-        "rotation": 0
+        "rotation": 0,
+        // for type regularShape
+        "rsRadius": 10,
+        "rsRadius2": undefined,
+        "rsPoints": 3,
+        "rsFillColor": [0, 153, 255, 1],
+        "rsStrokeColor": [0, 0, 0, 1],
+        "rsStrokeWidth": 5,
+        "rsAngle": 0,
+        "rsScale": undefined
     },
 
     initialize: function (feature, styles, isClustered) {
@@ -113,6 +122,9 @@ const PointStyleModel = StyleModel.extend(/** @lends PointStyleModel.prototype *
         }
         else if (type === "interval") {
             return this.createIntervalPointStyle();
+        }
+        else if (type === "regularshape") {
+            return this.createRegularShapeStyle();
         }
 
         return new Style();
@@ -288,6 +300,41 @@ const PointStyleModel = StyleModel.extend(/** @lends PointStyleModel.prototype *
             svgPath = styleScalingShape === "CIRCLE_BAR" ? this.createIntervalCircleBar(feature) : "";
 
         return this.createSVGStyle(svgPath);
+    },
+
+    /**
+     * Creates RegularShape Style.
+     * all features get same shape styöe.
+     * @returns {ol/style} - The created RegularShape style.
+     */
+    createRegularShapeStyle: function () {
+        const radius = parseFloat(this.get("rsRadius"), 10),
+            radius2 = parseFloat(this.get("rsRadius2"), 10),
+            points = parseFloat(this.get("rsPoints"), 10),
+            angle = parseFloat(this.get("rsAngle"), 10),
+            rotation = this.getRotationValue(this.get("rotation")),
+            scale = this.get("rsScale"),
+            fillcolor = this.returnColor(this.get("rsFillColor"), "rgb"),
+            strokecolor = this.returnColor(this.get("rsStrokeColor"), "rgb"),
+            strokewidth = parseFloat(this.get("rsStrokeWidth"), 10);
+
+        return new Style({
+            image: new RegularShape({
+                radius: radius,
+                radius2: radius2,
+                points: points,
+                scale: scale,
+                rotation: rotation,
+                angle: angle,
+                fill: new Fill({
+                    color: fillcolor
+                }),
+                stroke: new Stroke({
+                    color: strokecolor,
+                    width: strokewidth
+                })
+            })
+        });
     },
 
     /**