+++ src/core/layers/sta.js
@@ -810,6 +810,7 @@ STALayer.prototype.createPropertiesOfDatastreamsHelper = function (dataStreams, 
         const dataStreamId = String(dataStream["@iot.id"]),
             dataStreamName = dataStream.name,
             dataStreamValue = Array.isArray(dataStream.Observations) ? dataStream.Observations[0]?.result : "",
+            dataStreamUnit = dataStream.unitOfMeasurement?.name,
             key = "dataStream_" + dataStreamId + "_" + dataStreamName;
         let phenomenonTime = Array.isArray(dataStream.Observations) ? dataStream.Observations[0]?.phenomenonTime : "";
 
@@ -829,6 +830,13 @@ STALayer.prototype.createPropertiesOfDatastreamsHelper = function (dataStreams, 
             properties[key] = noDataValue;
             properties[key + "_phenomenonTime"] = noDataValue;
             properties.dataStreamValue.push(noDataValue);
+        }
+
+        if (typeof dataStreamUnit !== "undefined" && typeof this.get("rotationUnit") !== "undefined" && dataStreamUnit === this.get("rotationUnit")) {
+            properties.rotation = {
+                isDegree: true,
+                value: typeof dataStreamValue !== "undefined" ? dataStreamValue : 0
+            };
         }
     });
 
@@ -1470,6 +1478,13 @@ STALayer.prototype.updateFeatureProperties = function (feature, dataStreamId, re
     feature.set("dataStream_" + dataStreamId + "_" + dataStreamName + "_phenomenonTime", phenomenonTime, true);
     feature.set("dataStreamValue", this.replaceValueInPipedProperty(feature, "dataStreamValue", dataStreamId, preparedResult));
     feature.set("dataStreamPhenomenonTime", this.replaceValueInPipedProperty(feature, "dataStreamPhenomenonTime", dataStreamId, phenomenonTime));
+
+    if (typeof feature.get("rotation") !== "undefined" && typeof preparedResult !== "undefined") {
+        feature.set("rotation", {
+            isDegree: true,
+            value: preparedResult
+        });
+    }
 
     if (typeof funcChangeFeatureGFI === "function") {
         funcChangeFeatureGFI(feature);